declare function as float, not double.
SCCS-vsn: usr.bin/f77/libF77/r_acos.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_asin.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_atan.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_atn2.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_cos.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_cosh.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_exp.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_log.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_sin.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_sinh.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_sqrt.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_tan.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_tanh.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_abs.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_imag.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_nint.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_int.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_sign.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_lg10.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_dim.c 5.2
SCCS-vsn: usr.bin/f77/libF77/r_mod.c 5.2

+++ usr/src/usr.bin/f77/libF77/r_abs.c
@@ -3,10 +3,10 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_abs.c	5.1	%G%
+ *	@(#)r_abs.c	5.2	%G%
  */
 
-double r_abs(x)
+float r_abs(x)
 float *x;
 {
 if(*x >= 0)
